Moved the huge scene test from SceneBuilderTests.cs to MeltyPLayerTests.cs.
