+++ tests/SharpGLTF.ThirdParty.Tests/SharpGLTF.ThirdParty.Tests.csproj
@@ -18,7 +18,8 @@
   <ItemGroup>
     <PackageReference Include="NUnit" Version="3.13.3" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.3.1" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.1" />    
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.1" />
+    <PackageReference Include="System.Drawing.Common" Version="6.0.0" />    
   </ItemGroup>
 
   <ItemGroup>
